Workaround Next.js bundler bug
This prevents the following error, which only appears in the browser
console in production (i.e. it does not occur with `npm run dev`):

    Error: Minified React error #306; visit https://react.dev/errors/306?args[]=undefined&args[]= for the full message or use the non-minified dev environment for full errors and additional helpful warnings.

According to the linked page, the full text of the error is:

    Element type is invalid. Received a promise that resolves to: undefined. Lazy element type must resolve to a class or function.

However, the actual issue is a bug in the Next.js bundler that omits
client components from the server JavaScript bundle if they are only
referenced by server actions.

Thus, the workaround is to manually reference the relevant client
component (in this case, `<Waterfalling>`) from server code other than
the server action.